SRC/Frontend/StreamLit/requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
apispec==6.6.1
apispec-webframeworks==1.1.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
attrs==24.3.0
Automat==24.8.1
autopep8==2.3.2
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boto3==1.34.113
botocore==1.34.162
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.4.1
fonttools==4.55.6
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
git-filter-repo==2.47.0
gitdb==4.0.12
gitignore_parser==0.1.11
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
kthread==0.2.3
lxml==5.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.2
openai==1.65.2
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
premailer==3.10.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.7.2
PyMySQL==1.1.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-slugify==8.0.4
python-socketio==5.12.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
selenium==4.28.1
simple-websocket==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.30
st-pages==1.0.1
stack-data==0.6.3
streamlit==1.43.1
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
taipy==4.0.2
taipy-common==4.0.2
taipy-core==4.0.2
taipy-gui==4.0.2
taipy-rest==4.0.2
taipy-templates==4.0.2
tenacity==9.0.0
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.7.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yagmail==0.15.293
yapf==0.43.0
yarg==0.1.9
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
